6/5/2015 - Program Improvement and/or Correction
I discovered that the speed & counter variables were not clearing properly.  This was corrected by adding counter = 0 and speed = 0 to the primary WHILE loop.  This error resulted in a miscalculation of the speed.

The LCD code was also relocated to speed the processing of the counter and reduce program lag time.